YBE-1582 Make libpng prefixable to avoid conflicts with system version
The libpng library proper has added the ability to add custom prefixes for its public APIs
a while ago, and it's actually quite easy to use (basically a one-liner, specifying prefix
as a PNG_PREFIX argument - which is also in the accompanying PR YOU-i-Labs/YouiPlatform#17846 (review)).

Yet, when I tested our libpng fork, it turned out that while specifying the prefix
via the configure script worked fine, the cmake alternative didn't work at all.
I tested it with the pure upstream and it all worked in both modes,
off the tip and with the version we forked from as well.

After some bisecting, I arrived at commit 25399ab which both hunterified the libpng for us
AND broke its prefixing functionality.

The purpose of this PR is to restore the ability to add prefixes while keeping the hunter integration.